+++ monkey/common/common_consts/telem_categories.py
@@ -2,7 +2,6 @@ class TelemCategoryEnum:
     EXPLOIT = "exploit"
     POST_BREACH = "post_breach"
     SCAN = "scan"
-    SCOUTSUITE = "scoutsuite"
     STATE = "state"
     SYSTEM_INFO = "system_info"
     TRACE = "trace"