+++ monkey/monkey_island/cc/resources/zero_trust/zero_trust_report.py
@@ -1,7 +1,7 @@
 import http.client
 
 import flask_restful
-from flask import Response, jsonify
+from flask import jsonify
 
 from monkey_island.cc.resources.auth.auth import jwt_required
 from monkey_island.cc.services.zero_trust.zero_trust_report.finding_service import FindingService
@@ -9,14 +9,10 @@ from monkey_island.cc.services.zero_trust.zero_trust_report.pillar_service impor
 from monkey_island.cc.services.zero_trust.zero_trust_report.principle_service import (
     PrincipleService,
 )
-from monkey_island.cc.services.zero_trust.zero_trust_report.scoutsuite_raw_data_service import (
-    ScoutSuiteRawDataService,
-)
 
 REPORT_DATA_PILLARS = "pillars"
 REPORT_DATA_FINDINGS = "findings"
 REPORT_DATA_PRINCIPLES_STATUS = "principles"
-REPORT_DATA_SCOUTSUITE = "scoutsuite"
 
 
 class ZeroTrustReport(flask_restful.Resource):
@@ -28,10 +24,5 @@ class ZeroTrustReport(flask_restful.Resource):
             return jsonify(PrincipleService.get_principles_status())
         elif report_data == REPORT_DATA_FINDINGS:
             return jsonify(FindingService.get_all_findings_for_ui())
-        elif report_data == REPORT_DATA_SCOUTSUITE:
-            # Raw ScoutSuite data is already solved as json, no need to jsonify
-            return Response(
-                ScoutSuiteRawDataService.get_scoutsuite_data_json(), mimetype="application/json"
-            )
 
         flask_restful.abort(http.client.NOT_FOUND)