+++ monkey/monkey_island/cc/services/zero_trust/scoutsuite/consts/rule_names/rule_name_enum.py
@@ -1,5 +1,0 @@
-from enum import Enum
-
-
-class RuleNameEnum(Enum):
-    pass